requirements.txt
# Core testing framework
playwright>=1.42.0
pytest>=7.4.0
pytest-html>=4.1.1
pytest-json-report>=1.5.0
beautifulsoup4>=4.12.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# AI/LLM integration
groq>=0.4.0

# HTTP client
requests>=2.31.0

